requirements.txt
bcrypt==3.1.1
cffi==1.8.3
Django==1.10.2
Pillow==3.4.1
psycopg2==2.6.2
pycparser==2.14
requests==2.11.1
six==1.10.0
whitenoise==3.2.2
